refactored xmtp feedback flow
